perf: bf16 model weights + verified batch configs for 1.7B
Changed LLM torch_dtype from float32 to bfloat16, saving ~3.5GB GPU
memory (27.7→20.8GB fixed). Verified on A100 80GB:
- batch_size=16: 100% GPU util, 8 samp/s, 0 OOMs, 57GB total
- batch_size=24: ~15% OOM rate on long-video batches — too aggressive
Updated config comments with verified memory/throughput numbers.
Added SmolVLM2 comparison scores to benchmark results.

+++ fVLM/release/model/foveated_vlm.py
@@ -64,7 +64,7 @@ class FoveatedVLM(nn.Module):
 
         # ---- Language model ----
         self.llm = AutoModelForCausalLM.from_pretrained(
-            llm_name, attn_implementation="sdpa", torch_dtype=torch.float32,
+            llm_name, attn_implementation="sdpa", torch_dtype=torch.bfloat16,
         )
         self.llm.config.use_cache = False  # training default; overridden per-method
         llm_dim = self.llm.config.hidden_size
